checkpoint: before fixing validation refactoring test issues

+++ src/main/java/br/unifor/costify/application/dto/entity/IngredientDto.java
@@ -16,7 +16,7 @@ public record IngredientDto(
         ingredient.getId().getValue(),
         ingredient.getName(),
         ingredient.getPackageQuantity(),
-        ingredient.getPackagePrice(),
+        ingredient.getPackagePrice().doubleValue(),
         ingredient.getPackageUnit(),
         ingredient.getUnitCost());
   }
